titanic_project/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
PyYAML>=6.0
